// src/models/card.model.ts
// A player's statistical profile
export interface CardModel {
  xGF: number;
  xGA: number;
  primaryAssistsPP: number;
  primaryAssistsEV: number;
  TOI: number;
  shortHandedTOI: number;
  powerPlayTOI: number;
  gsae: number;
  sh_xGA: number;
  goals: number;
  pp: number;
  low_danger: number;
  med_danger: number;
  high_danger: number;
  ev: number;
  starts: number;
  shootout: number;
  med_danger_freq: number;
  high_danger_freq: number;
  pk: number;
}